feat: ws using system proxy config
Reads the system proxy settings and if that is non nil, sets the
socksproxy property of the ws socket. Notice that the system proxy
config could contain other proxy types, like HTTP proxies, but they
seem to be ignored.

The motivation for doing this is using a SOCKS proxy inside the docker
network exposed to the guest network. This allows the guest system to
use docker dns and private IPs without exposing local ports. This give
us more options to test the system, as we could use a remote docker
deployment.